requirements.txt
# Patient EHR System - Python Dependencies
# Alternative to pyproject.toml for traditional pip installations

# Core Django
django>=5.2.7
djangorestframework>=3.16.1
django-cors-headers>=4.9.0

